Added more accurate node counter, changes metrics to 64bit integers

+++ src/Metrics.h
@@ -8,7 +8,7 @@
 #include <unordered_map>
 #include <string>
 
-template<int ID, typename T = int>
+template<int ID, typename T = int64_t>
 class Metric {
  public:
     static void inc(T n = 1) {
